requirements.txt
coreapi==2.0.9
Django==1.8.2
django-cors-headers==1.3.1
django-rest-swagger==2.1.0
djangorestframework==3.5.3
itypes==1.1.0
openapi-codec==1.1.7
psycopg2==2.6.2
requests==2.12.1
simplejson==3.10.0
uritemplate==3.0.0